[FEAT] Add PlayerBumpkin component
